Check if libMonoPosixHelper.so exists before cp
Testing stuff is hard

+++ default.nix
@@ -121,7 +121,9 @@ in buildDotnetPackage rec {
       $out/lib/dotnet/Everest/libMonoPosixHelper.dylib.dSYM/Contents/Info.plist \
       $out/lib/dotnet/Everest/lib64/* \
       $out/lib/dotnet/Everest/
-    cp ${mono}/lib/libMonoPosixHelper.so $out/lib/dotnet/Everest
+    if [ -f "${mono}/lib/libMonoPosixHelper.so"]; then
+      cp ${mono}/lib/libMonoPosixHelper.so $out/lib/dotnet/Everest
+    fi
     rm -r $out/lib/dotnet/Everest/lib64 $out/lib/dotnet/Everest/libMonoPosixHelper.dylib.dSYM
     sed -i "2i chmod -R u+w ." $out/bin/miniinstaller
     sed -i "2i cp -r $out/lib/dotnet/Everest/* "'$1' $out/bin/miniinstaller
